project4/requirements.txt
bs4==0.0.1
channels==2.2
channels-redis==2.1.1
Django==2.2
django-crontab==0.7.1
pycodestyle==2.3.1
pytz==2017.2
redis==3.3.6
python-dateutil==2.8.0
